#132 Quan ly Order tai Admin

+++ src/main/java/vn/hoidanit/laptopshop/domain/Order.java
@@ -18,7 +18,7 @@ public class Order {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
 
-    private long totalPrice;
+    private double totalPrice;
 
     private String receiverName;
 
@@ -49,11 +49,11 @@ public class Order {
         this.id = id;
     }
 
-    public long getTotalPrice() {
+    public double getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(long totalPrice) {
+    public void setTotalPrice(double totalPrice) {
         this.totalPrice = totalPrice;
     }
